refactor: Removal of old cobra structures
Restructure of code and removal of old elements.
