+++ go.mod
@@ -6,7 +6,6 @@ require (
 	github.com/julienschmidt/httprouter v1.2.1-0.20181021223831-26a05976f9bf
 	github.com/mitchellh/go-homedir v1.0.0
 	github.com/russross/blackfriday/v2 v2.1.0
-	github.com/spf13/cobra v0.0.4-0.20181021141114-fe5e611709b0
 	github.com/spf13/viper v1.2.1
 	golang.org/x/sys v0.0.0-20181122145206-62eef0e2fa9b
 )
@@ -15,7 +14,6 @@ require (
 	github.com/BurntSushi/toml v1.2.0 // indirect
 	github.com/fsnotify/fsnotify v1.4.7 // indirect
 	github.com/hashicorp/hcl v1.0.0 // indirect
-	github.com/inconshreveable/mousetrap v1.0.1 // indirect
 	github.com/magiconair/properties v1.8.0 // indirect
 	github.com/mitchellh/mapstructure v1.0.0 // indirect
 	github.com/pelletier/go-toml v1.2.0 // indirect
